llvm-toolchain: Invoke llvm assembler and linker instead.
Instead of invoking clang to do the linking (which is the default
behavior of CMake), invoke ld.lld directly so that cross-compiling
is easier. Likewise, instead of invoking the system assembler (as),
invoke the LLVM builtin assembler.

This is necessary when the system assembler and linker is for a
different architecture, such as when cross-compiling for x86 on an
Android aarch64 system.